Working example of SDL_AddEventWatch hacked together.
