+++ examples/hello-world/sdlConfig.ts
@@ -3,6 +3,7 @@ import { SDL } from "SDL_ts";
 // This file contains the list of functions that are used in the project.
 
 export const SDL_FUNCTIONS = [
+  SDL.AddEventWatch,
   SDL.CreateWindow,
   SDL.Delay,
   SDL.DestroyWindow,
